add gff-prot option and fix deprecations

+++ cctyper/castyping.py
@@ -33,9 +33,9 @@ class Typer(object):
         tmpX['Hmm'] = [re.sub("_.*","",x) for x in tmpX['Hmm']]
         tmpX.drop_duplicates('Hmm', inplace=True)
 
-        start = list(tmp['start'])
-        end = list(tmp['end'])
-
+        start = list(tmp['start'].astype(int))
+        end = list(tmp['end'].astype(int))
+        
         if operon in self.circ_operons:
             gene_end = np.argmax(np.diff(list(tmp['Pos'])))
             start_operon = start[gene_end+1]
